+++ changelog.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Setup global application design.
 * Setup authentication from `leto-modelizer`.
 * Setup Users page
+* Setup Libraries page